blog/pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
django = "^6.0.1"
ipython = "^9.9.0"
django-extensions = "^4.1"
django-summernote = "^0.8.20.0"
pillow = "^12.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
